Creado Enemigo Jefe

+++ test/core/src/entidades/Disparo.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.graphics.Texture;
 
 public class Disparo extends Entidad{
 	
-	public Disparo(float x, float y , float altura, float anchura, int vida, int velocidadX, int velocidadY, Texture textura) {
+	public Disparo(float x, float y , float altura, float anchura, int vida, float velocidadX, float velocidadY, Texture textura) {
 		super(x, y, altura, anchura, vida, velocidadX, velocidadY, textura);
 	}
